The customer review dataset of whatsapp
WhatsApp is a social media platform. I have downloaded the customer review dataset from Kaggle. I want to analyse the data by preprocessing it using pandas of pythons and then the extraction of data will be done using MySQL . and then the excel files will be analysed using Tableau. By this analysis, We will come to know the best version of WhatsApp. Then we can include the features in Gossip Hub similar to the best version of the WhatsApp.