WhatsApp Customer review preprocessed file
This is the file that I have preprocessed using Pandas library of Python. This file is important for further analysis. BY analysing the file, we get to know the best version of WhatsApp. Then we can copy the features of the best version of WhatsApp to Gossip Hub. Please refer to the final word document to see the final report of the project.